// WADNR.Web/src/app/shared/generated/model/project-organizations-step.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectOrganizationStepItem } from './project-organization-step-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectOrganizationsStep { 
    ProjectID?: number;
    Organizations?: Array<ProjectOrganizationStepItem> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectOrganizationsStepForm { 
    ProjectID?: FormControl<number>;
    Organizations?: FormControl<Array<ProjectOrganizationStepItem>>;
}

export class ProjectOrganizationsStepFormControls { 
    public static ProjectID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static Organizations = (value: FormControlState<Array<ProjectOrganizationStepItem>> | Array<ProjectOrganizationStepItem> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectOrganizationStepItem>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}