// WADNR.Web/src/app/shared/generated/model/project-organizations-step-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectOrganizationRequestItem } from './project-organization-request-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectOrganizationsStepRequest { 
    Organizations?: Array<ProjectOrganizationRequestItem> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectOrganizationsStepRequestForm { 
    Organizations?: FormControl<Array<ProjectOrganizationRequestItem>>;
}

export class ProjectOrganizationsStepRequestFormControls { 
    public static Organizations = (value: FormControlState<Array<ProjectOrganizationRequestItem>> | Array<ProjectOrganizationRequestItem> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectOrganizationRequestItem>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}